intra candidates
rd=3,4 or rd=5,6 should be the same, except this one.
just want to test what this does.
there's no check on the value but I think MAX_RD_INTRA_MODES = 16 is a safe maximum (candCostList and rdModeList are of this size)
however the CLI doesn't crash until 27 LOL I don't know why