test fix unnecessary intra blocks in multiple slices
Thanks rwill for pointing this out.

The constants cause MVs to be restricted to >12 or <-16 in slice boundaries.

example:
    minY: -2036, maxY: -16, or
    minY: 12, maxY: 1776

Which means: the block with little to none vertical motions will have high change to be mistakenly encoded as Intra blocks, or have terrible prediction.

The constants were introduced in a8d7d3d as 2*4 and 3*4, later changed to 3*4 and 4*4 in eaeeab5... but I don't understand why.
Judging by the second commit they seem to related to deterministic, but as I tested it just now... they still aren't.
So I think there's no problem to just remove them.

Frame-threads=1 seemingly can help with determinism so I forced it with slices, also the restriction is applied when "m_bFrameParallel" is true, so forcing it alone would be enough to "fix" the bug.
(although it's not like that it's very important after the constants are removed)